fix: classroomPage으로 소켓 이동 quest 컴포넌트들에게 있던 소켓코드를 classroomPage에서 다룰수 있도록 옮김
